At the time of writing, Safari support is just terrible
First of all, Safari needs the permission to access microphone even to
process <audio> elements. Even with the permission granted, there are
clicks and pops and generally sluggish experience. Microphone source
seems to work alright, however.

There are multiple known bugs in createMediaElementSource()
implementation of Safari; maybe when these are fixed it will work
properly:

https://bugs.webkit.org/buglist.cgi?quicksearch=createMediaElementSource

In the meanwhile, the workaround is to ask for microphone permission and
hope for the best:

https://github.com/mdn/webaudio-examples/issues/56